fix: arrange output parser classes
Put all the output parser classes in one place.
Remove redundant output parser classes.
Reuse output parser classes with the same functionality.
