--- agentverse/tasks/tasksolving/pythoncalculator/output_parser.py
@@ -1,19 +1,0 @@
-from __future__ import annotations
-
-import re
-from typing import Union, List, Tuple
-
-from agentverse.utils import AgentAction, AgentFinish, AgentCriticism
-
-from agentverse.parser import OutputParserError, output_parser_registry, OutputParser
-from agentverse.llms import LLMResult
-from agentverse.logging import get_logger
-
-
-logger = get_logger()
-
-
-@output_parser_registry.register("dummy")
-class PipelinePythoncalculatorParser(OutputParser):
-    def parse(self, output: LLMResult) -> Union[AgentAction, AgentFinish]:
-        return AgentFinish({"output": output.content}, output.content)
